requirements.txt
discord.py==1.5.1
giphy_client==1.0.0
rockset==0.7.42